requirements.txt
click >= 8.0.4 # library to create CLI
request >= 2.27.1